chore(docs): add comprehensive source of truth document for project
This commit introduces a detailed "Source of Truth" document aimed at providing a
master reference for the project. The document serves multiple purposes, including
a guide for resume building, interview preparation, and as a quick-reference
manual. It outlines the project's architecture, responsibilities, tech stack,
and operational guidelines, ensuring that all team members have access to
consistent and comprehensive information. This addition enhances knowledge
sharing and onboarding processes within the team.

+++ services/transaction-processor/src/main/java/com/tradestream/transaction_processor/consumer/TradeExecutedEvent.java
@@ -12,7 +12,7 @@ public class TradeExecutedEvent {
     private UUID buyOrderId;
     private UUID sellOrderId;
     private String ticker;
-    private BigDecimal price;
-    private int quantity;
-    private Instant timestamp; // a unix epoch
+    private BigDecimal price;     // keep BigDecimal
+    private BigDecimal quantity;  // <— was int
+    private Instant timestamp;    // ISO-8601 instant (producer side decides format)
 }
